requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy
tensorflow
python-box==6.0.2
pyYAML
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
notebook
sqlalchemy
pymysql
-e .